+++ .github/workflows/docker-builds.yml
@@ -16,7 +16,11 @@ jobs:
         python_version: [3.6, 3.7, 3.8]
         pytorch_version: [1.3, 1.4, 1.5]
     steps:
-      - uses: actions/checkout@v2
+      - name: Checkout
+        uses: actions/checkout@v2
+      - uses: actions/setup-python@v2
+        with:
+          python-version: 3.7
 
       - name: Publish Master to Docker
         # publish master
