+++ benchmarks/test_parity.py
@@ -4,7 +4,7 @@ import numpy as np
 import pytest
 import torch
 
-import tests.base.utils as tutils
+import tests.base.develop_utils as tutils
 from pytorch_lightning import Trainer, seed_everything
 from tests.base.models import ParityModuleRNN, ParityModuleMNIST
