+++ pytorch_lightning/trainer/trainer.py
@@ -52,7 +52,7 @@ else:
 
 try:
     import horovod.torch as hvd
-except ImportError:
+except (ModuleNotFoundError, ImportError):
     HOROVOD_AVAILABLE = False
 else:
     HOROVOD_AVAILABLE = True
@@ -255,7 +255,7 @@ class Trainer(
 
                     Use `row_log_interval` instead. Will remove 0.9.0.
 
-            distributed_backend: The distributed backend to use (dp, ddp, ddp2, ddp_spawn)
+            distributed_backend: The distributed backend to use (dp, ddp, ddp2, ddp_spawn, ddp_cpu)
 
             use_amp:
                 .. warning:: .. deprecated:: 0.7.0
@@ -885,7 +885,7 @@ class Trainer(
                 task = int(os.environ['LOCAL_RANK'])
                 self.ddp_train(task, model)
 
-            elif self.distributed_backend == 'cpu_ddp':
+            elif self.distributed_backend == 'ddp_cpu':
                 self.set_random_port()
                 self.model = model
                 mp.spawn(self.ddp_train, nprocs=self.num_processes, args=(model,))
