+++ tests/base/model_utilities.py
@@ -4,7 +4,6 @@ from tests.base.datasets import TrialMNIST
 
 
 class ModelTemplateData:
-    hparams: ...
 
     def dataloader(self, train: bool, num_samples: int = 100):
         dataset = TrialMNIST(root=self.data_root, train=train, num_samples=num_samples, download=True)
