+++ tests/callbacks/test_callbacks.py
@@ -3,7 +3,7 @@ from pathlib import Path
 import pytest
 import torch
 
-import tests.base.utils as tutils
+import tests.base.develop_utils as tutils
 from pytorch_lightning import Callback
 from pytorch_lightning import Trainer, LightningModule
 from pytorch_lightning.callbacks import EarlyStopping, ModelCheckpoint
@@ -312,11 +312,12 @@ def test_model_checkpoint_with_non_string_input(tmpdir, save_top_k):
 
     checkpoint = ModelCheckpoint(filepath=None, save_top_k=save_top_k)
 
-    trainer = Trainer(default_root_dir=tmpdir,
-                      checkpoint_callback=checkpoint,
-                      overfit_batches=0.20,
-                      max_epochs=2
-                      )
+    trainer = Trainer(
+        default_root_dir=tmpdir,
+        checkpoint_callback=checkpoint,
+        overfit_batches=0.20,
+        max_epochs=2,
+    )
     trainer.fit(model)
 
     # These should be different if the dirpath has be overridden
@@ -337,7 +338,7 @@ def test_model_checkpoint_path(tmpdir, logger_version, expected):
         default_root_dir=tmpdir,
         overfit_batches=0.2,
         max_epochs=2,
-        logger=logger
+        logger=logger,
     )
     trainer.fit(model)
