+++ tests/loggers/test_all.py
@@ -3,7 +3,7 @@ import pickle
 
 import pytest
 
-import tests.base.utils as tutils
+import tests.base.develop_utils as tutils
 from pytorch_lightning import Trainer
 from pytorch_lightning.loggers import (
     TensorBoardLogger, MLFlowLogger, NeptuneLogger, TestTubeLogger, CometLogger)
@@ -90,7 +90,7 @@ def test_loggers_pickle(tmpdir, monkeypatch, logger_class):
 
     trainer = Trainer(
         max_epochs=1,
-        logger=logger
+        logger=logger,
     )
     pkl_bytes = pickle.dumps(trainer)
 
@@ -110,7 +110,7 @@ def test_logger_reset_correctly(tmpdir, extra_params):
 
     trainer = Trainer(
         default_root_dir=tmpdir,
-        **extra_params
+        **extra_params,
     )
     logger1 = trainer.logger
     trainer.fit(model)
