+++ tests/loggers/test_neptune.py
@@ -83,7 +83,7 @@ def test_neptune_leave_open_experiment_after_fit(tmpdir):
             default_root_dir=tmpdir,
             max_epochs=1,
             limit_train_batches=0.05,
-            logger=logger
+            logger=logger,
         )
         trainer.fit(model)
         return logger
