+++ tests/models/data/horovod/train_default_model.py
@@ -21,7 +21,11 @@ import json
 import os
 import sys
 
-import horovod.torch as hvd
+
+try:
+    import horovod.torch as hvd
+except (ModuleNotFoundError, ImportError):
+    print('You requested to import Horovod which is missing or not supported for your OS.')
 
 PATH_HERE = os.path.abspath(os.path.dirname(__file__))
 PATH_ROOT = os.path.join(PATH_HERE, '..', '..', '..', '..')
@@ -30,7 +34,8 @@ sys.path.insert(0, os.path.abspath(PATH_ROOT))
 from pytorch_lightning import Trainer  # noqa: E402
 from pytorch_lightning.callbacks import ModelCheckpoint  # noqa: E402
 from tests.base import EvalModelTemplate  # noqa: E402
-from tests.base.utils import set_random_master_port, run_model_test  # noqa: E402
+from tests.base.develop_pipelines import run_model_test  # noqa: E402
+from tests.base.develop_utils import set_random_master_port  # noqa: E402
 
 
 parser = argparse.ArgumentParser()
