+++ tests/models/test_cpu.py
@@ -5,7 +5,8 @@ import pytest
 import torch
 from packaging.version import parse as version_parse
 
-import tests.base.utils as tutils
+import tests.base.develop_pipelines as tpipes
+import tests.base.develop_utils as tutils
 from pytorch_lightning import Trainer
 from pytorch_lightning.callbacks import EarlyStopping
 from pytorch_lightning.callbacks import ModelCheckpoint
@@ -27,7 +28,7 @@ def test_cpu_slurm_save_load(tmpdir):
         logger=logger,
         limit_train_batches=0.2,
         limit_val_batches=0.2,
-        checkpoint_callback=ModelCheckpoint(tmpdir)
+        checkpoint_callback=ModelCheckpoint(tmpdir),
     )
     result = trainer.fit(model)
     real_global_step = trainer.global_step
@@ -97,7 +98,7 @@ def test_early_stopping_cpu_model(tmpdir):
     )
 
     model = EvalModelTemplate()
-    tutils.run_model_test(trainer_options, model, on_gpu=False)
+    tpipes.run_model_test(trainer_options, model, on_gpu=False)
 
     # test freeze on cpu
     model.freeze()
@@ -122,11 +123,11 @@ def test_multi_cpu_model_ddp(tmpdir):
         limit_val_batches=0.2,
         gpus=None,
         num_processes=2,
-        distributed_backend='ddp_cpu'
+        distributed_backend='ddp_cpu',
     )
 
     model = EvalModelTemplate()
-    tutils.run_model_test(trainer_options, model, on_gpu=False)
+    tpipes.run_model_test(trainer_options, model, on_gpu=False)
 
 
 def test_lbfgs_cpu_model(tmpdir):
@@ -145,7 +146,7 @@ def test_lbfgs_cpu_model(tmpdir):
                    learning_rate=0.004)
     model = EvalModelTemplate(**hparams)
     model.configure_optimizers = model.configure_optimizers__lbfgs
-    tutils.run_model_test_without_loggers(trainer_options, model, min_acc=0.25)
+    tpipes.run_model_test_without_loggers(trainer_options, model, min_acc=0.25)
 
 
 def test_default_logger_callbacks_cpu_model(tmpdir):
@@ -161,7 +162,7 @@ def test_default_logger_callbacks_cpu_model(tmpdir):
     )
 
     model = EvalModelTemplate()
-    tutils.run_model_test_without_loggers(trainer_options, model)
+    tpipes.run_model_test_without_loggers(trainer_options, model)
 
     # test freeze on cpu
     model.freeze()
@@ -187,7 +188,7 @@ def test_running_test_after_fitting(tmpdir):
         limit_val_batches=0.2,
         limit_test_batches=0.2,
         checkpoint_callback=checkpoint,
-        logger=logger
+        logger=logger,
     )
     result = trainer.fit(model)
 
@@ -218,7 +219,7 @@ def test_running_test_no_val(tmpdir):
         limit_test_batches=0.2,
         checkpoint_callback=checkpoint,
         logger=logger,
-        early_stop_callback=False
+        early_stop_callback=False,
     )
     result = trainer.fit(model)
 
@@ -259,7 +260,7 @@ def test_cpu_model(tmpdir):
 
     model = EvalModelTemplate()
 
-    tutils.run_model_test(trainer_options, model, on_gpu=False)
+    tpipes.run_model_test(trainer_options, model, on_gpu=False)
 
 
 def test_all_features_cpu_model(tmpdir):
@@ -277,7 +278,7 @@ def test_all_features_cpu_model(tmpdir):
     )
 
     model = EvalModelTemplate()
-    tutils.run_model_test(trainer_options, model, on_gpu=False)
+    tpipes.run_model_test(trainer_options, model, on_gpu=False)
 
 
 def test_tbptt_cpu_model(tmpdir):
@@ -344,7 +345,7 @@ def test_tbptt_cpu_model(tmpdir):
         truncated_bptt_steps=truncated_bptt_steps,
         limit_val_batches=0,
         weights_summary=None,
-        early_stop_callback=False
+        early_stop_callback=False,
     )
     result = trainer.fit(model)
