+++ tests/models/test_grad_norm.py
@@ -5,7 +5,7 @@ from pytorch_lightning import Trainer
 from pytorch_lightning.loggers import LightningLoggerBase
 from pytorch_lightning.utilities import rank_zero_only
 from tests.base import EvalModelTemplate
-from tests.base.utils import reset_seed
+from tests.base.develop_utils import reset_seed
 
 
 class OnlyMetricsListLogger(LightningLoggerBase):
