+++ tests/trainer/test_optimizers.py
@@ -17,7 +17,7 @@ def test_optimizer_with_scheduling(tmpdir):
         default_root_dir=tmpdir,
         max_epochs=1,
         limit_val_batches=0.1,
-        limit_train_batches=0.2
+        limit_train_batches=0.2,
     )
     results = trainer.fit(model)
     assert results == 1
@@ -48,7 +48,7 @@ def test_multi_optimizer_with_scheduling(tmpdir):
         default_root_dir=tmpdir,
         max_epochs=1,
         limit_val_batches=0.1,
-        limit_train_batches=0.2
+        limit_train_batches=0.2,
     )
     results = trainer.fit(model)
     assert results == 1
@@ -83,7 +83,7 @@ def test_multi_optimizer_with_scheduling_stepping(tmpdir):
         default_root_dir=tmpdir,
         max_epochs=1,
         limit_val_batches=0.1,
-        limit_train_batches=0.2
+        limit_train_batches=0.2,
     )
     results = trainer.fit(model)
     assert results == 1
@@ -122,7 +122,7 @@ def test_reduce_lr_on_plateau_scheduling(tmpdir):
         default_root_dir=tmpdir,
         max_epochs=1,
         limit_val_batches=0.1,
-        limit_train_batches=0.2
+        limit_train_batches=0.2,
     )
     results = trainer.fit(model)
     assert results == 1
@@ -212,7 +212,7 @@ def test_none_optimizer(tmpdir):
         default_root_dir=tmpdir,
         max_epochs=1,
         limit_val_batches=0.1,
-        limit_train_batches=0.2
+        limit_train_batches=0.2,
     )
     result = trainer.fit(model)
 
@@ -234,6 +234,9 @@ def test_configure_optimizer_from_dict(tmpdir):
     model = CurrentModel(hparams)
 
     # fit model
-    trainer = Trainer(default_root_dir=tmpdir, max_epochs=1)
+    trainer = Trainer(
+        default_root_dir=tmpdir,
+        max_epochs=1,
+    )
     result = trainer.fit(model)
     assert result == 1
