+++ tests/trainer/test_trainer_tricks.py
@@ -1,11 +1,11 @@
 import pytest
 import torch
+from torch.utils.data import RandomSampler, SequentialSampler, DataLoader
 
-import tests.base.utils as tutils
+import tests.base.develop_utils as tutils
 from pytorch_lightning import Trainer
 from pytorch_lightning.utilities.exceptions import MisconfigurationException
 from tests.base import EvalModelTemplate
-from torch.utils.data import RandomSampler, SequentialSampler, DataLoader
 
 
 def test_overfit_batch_limits(tmpdir):
